+++ lib/extended_email_reply_parser/parsers/github.rb
@@ -1,5 +1,6 @@
 module ExtendedEmailReplyParser
   class Parsers::Github < Parsers::Base
+    add_quote_header_regex '^On .* wrote:$'
 
     def parse
       EmailReplyParser.parse_reply text